student-center/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
psycopg2-binary
redis
celery
pytest
google-generativeai
google-genai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
bcrypt==3.2.0
httpx
requests
python-dotenv
python-multipart
aiofiles
email-validator
fastapi-mail
slowapi

# OCR Pipeline (Tesseract + TextToLatex Mamba)
torch
transformers
pymupdf
Pillow
pytesseract
